Only start game on game screen
Why:
Currently the game starts as soon as the app loads the welcome screen.
We would like for the game to only start after the user as pressed play
on the welcome screen and have navigated to the game board.

This commit:
Moves the GameProvider to wrap only the game screen and not the
AppNavigatorContainer. This will case the game context to only render
(and thus start the game) after the user has navigated to the game
screen.
